build: Remove laravel container
The laravel container was not useful because everything it writes is as root, so files cannot be edited from the host.
This caused commands like 'php artisan make' to not work as expected, and so it's useless to have a container
running if we end up using php commands from the host.