feat: Add profile relationship to model_has_role table
The model_has_role table, represented as ModelRolePivot model, now contains a
polymorphic relationship with profiles. This means that for a given User (model)
and Role, there may be a profile with extra information about the User in that role.

At the moment there isn't a explicit constraint that uniquely relates a Role with a
type of Profile, so this would need to be done at the business logic level.