feat: Add UserController for admin functionality
This includes the addition of CascadeSoftDelete, use to
propagate the SoftDelete of users by admins.
If this wouldn't be this way, then Products would still appear
as available even if the user doesn't exist anymore. Also, other pages wouldn't be able to display
profile information since they would be hard deleted.